Fixed a race condition issue in Asterisk applications.
The issue applies to all the apps: SynthAndRecog, MRCPRecog, and MRCPSynth, where an application may get stuck infinitely in speech_channel_open() having the following conditions met.

An application is getting executed, the establishment of a new synthesizer or recognizer session/channel fails due to a license issue or other error in the MRCP server, and as a consequence, the calling thread of the application may (not always) get stuck waiting for a response from the MRCP client. The response is actually signaled but not delivered/received due to a race in the function speech_channel_open().